+++ app/src/main/java/com/sopt/dive/navigation/BottomBar.kt
@@ -1,4 +1,4 @@
-package com.sopt.dive.ui.screen
+package com.sopt.dive.navigation
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.RowScope
@@ -10,7 +10,6 @@ import androidx.compose.material3.Icon
 import androidx.compose.material3.NavigationBar
 import androidx.compose.material3.NavigationBarItem
 import androidx.compose.material3.NavigationBarItemDefaults
-import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
@@ -20,31 +19,9 @@ import androidx.navigation.NavDestination
 import androidx.navigation.NavGraph.Companion.findStartDestination
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.currentBackStackEntryAsState
-import androidx.navigation.compose.rememberNavController
 import com.sopt.dive.data.route.Home
 import com.sopt.dive.data.route.My
 import com.sopt.dive.data.route.Search
-import com.sopt.dive.navigation.BottomBarItem
-import com.sopt.dive.navigation.MainNavHost
-
-@Composable
-fun MainScreen(
-) {
-    val navController = rememberNavController()
-
-    Scaffold (
-        bottomBar = {
-            BottomBar(
-                navController = navController
-            )
-        }
-    ){ innerPadding ->
-        MainNavHost(
-            navController = navController,
-            innerPadding = innerPadding
-        )
-    }
-}
 
 @Composable
 fun BottomBar(
