+++ app/src/main/java/com/sopt/dive/navigation/MainNavHost.kt
@@ -5,6 +5,11 @@ import androidx.compose.runtime.Composable
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import com.sopt.dive.data.route.Login
+import com.sopt.dive.ui.screen.home.navigation.homeNavGraph
+import com.sopt.dive.ui.screen.login.navigation.loginNavGraph
+import com.sopt.dive.ui.screen.my.navigation.myNavGraph
+import com.sopt.dive.ui.screen.search.navigation.searchNavGraph
+import com.sopt.dive.ui.screen.signup.navigation.signupNavGraph
 
 @Composable
 fun MainNavHost(
