+++ app/src/main/java/com/sopt/dive/ui/screen/my/navigation/MyNavigation.kt
@@ -1,10 +1,10 @@
-package com.sopt.dive.navigation
+package com.sopt.dive.ui.screen.my.navigation
 
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
 import com.sopt.dive.data.route.My
-import com.sopt.dive.ui.screen.MyScreen
+import com.sopt.dive.ui.screen.my.MyScreen
 
 fun NavGraphBuilder.myNavGraph(
     innerPadding: PaddingValues
