+++ app/src/main/java/com/sopt/dive/ui/screen/search/navigation/SearchNavigation.kt
@@ -1,10 +1,10 @@
-package com.sopt.dive.navigation
+package com.sopt.dive.ui.screen.search.navigation
 
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
 import com.sopt.dive.data.route.Search
-import com.sopt.dive.ui.screen.SearchScreen
+import com.sopt.dive.ui.screen.search.SearchScreen
 
 fun NavGraphBuilder.searchNavGraph(
     innerPadding: PaddingValues
